Cocoa Port: - Actually use -Ofast and -fvectorize in the Xcode 5 project. - Replace data type unsigned int with size_t where appropriate. - Do some minor code cleanup.

+++ desmume/src/cocoa/cocoa_input.mm
@@ -36,7 +36,7 @@
 		return self;
 	}
 	
-	for (unsigned int i = 0; i < DSControllerState_StatesCount; i++)
+	for (size_t i = 0; i < DSControllerState_StatesCount; i++)
 	{
 		controllerState[i] = false;
 	}
@@ -95,7 +95,7 @@
 	const NSInteger theMicMode = micMode;
 	static bool flushedStates[DSControllerState_StatesCount] = {0};
 	
-	for (unsigned int i = 0; i < DSControllerState_StatesCount; i++)
+	for (size_t i = 0; i < DSControllerState_StatesCount; i++)
 	{
 		flushedStates[i] = controllerState[i];
 	}
